Title: 배열의 특정 위치 찾기, Time: 139ms, Memory: 10MB, Status: Passed - CodeTree

+++ 241025/배열의 특정 위치 찾기/find-specific-location-fo-array.java
@@ -23,7 +23,7 @@ public class Main {
             }
         }
 
-        double average = sum3 / cnt;
+        double average = (double)sum3 / cnt;
 
         System.out.printf("%d %.1f", sum, average);
     }
